+++ src/main/java/org/sistemaBancario/servicios/serviciosCuentaBancaria/CuentaBancariaServicioImpl.java
@@ -1,11 +1,9 @@
-package org.sistemaBancario.servicios;
+package org.sistemaBancario.servicios.serviciosCuentaBancaria;
 
 import org.sistemaBancario.domain.Cliente;
 import org.sistemaBancario.domain.CuentaBancaria;
 import org.sistemaBancario.domain.CuentaCorriente;
 import org.sistemaBancario.domain.CuentaDeAhorro;
-
-import java.util.Scanner;
 
 public class CuentaBancariaServicioImpl implements CuentaBancariaServicio {
