+++ src/main/java/org/sistemaBancario/servicios/serviciosCuentaDeAhorro/CuentaDeAhorroServicioImpl.java
@@ -1,10 +1,7 @@
-package org.sistemaBancario.servicios;
+package org.sistemaBancario.servicios.serviciosCuentaDeAhorro;
 
 import org.sistemaBancario.domain.CuentaBancaria;
 import org.sistemaBancario.domain.CuentaDeAhorro;
-
-
-import java.util.Scanner;
 
 public class CuentaDeAhorroServicioImpl implements CuentaDeAhorroServicio {
     @Override
