+++ src/main/java/org/sistemaBancario/servicios/serviciosMenu/MenuServicio.java
@@ -1,4 +1,4 @@
-package org.sistemaBancario.servicios;
+package org.sistemaBancario.servicios.serviciosMenu;
 
 public interface MenuServicio {
     void desplegarMenu();
