+++ src/main/java/org/sistemaBancario/servicios/serviciosMenu/menuCuentaBancaria/MenuCuentaBancariaServicio.java
@@ -1,4 +1,4 @@
-package org.sistemaBancario.servicios;
+package org.sistemaBancario.servicios.serviciosMenu.menuCuentaBancaria;
 
 public interface MenuCuentaBancariaServicio {
     double lecturaMontoParaDepositar();
