+++ src/main/java/org/sistemaBancario/servicios/serviciosMenu/menuCuentaCorriente/MenuCuentaCorrienteServicio.java
@@ -1,4 +1,4 @@
-package org.sistemaBancario.servicios;
+package org.sistemaBancario.servicios.serviciosMenu.menuCuentaCorriente;
 
 import org.sistemaBancario.domain.CuentaBancaria;
