+++ src/main/java/org/sistemaBancario/servicios/serviciosMenu/menuCuentaDeAhorro/MenuCuentaDeAhorroServicio.java
@@ -1,4 +1,4 @@
-package org.sistemaBancario.servicios;
+package org.sistemaBancario.servicios.serviciosMenu.menuCuentaDeAhorro;
 
 public interface MenuCuentaDeAhorroServicio {
     double lecturaNuevaTasaDeIntereses();
